Improved script saving.
Each level now has its own script files.  For example, if I just beat
level 3 and went back to play level 1, the level 1 script would be
loaded, not the level3 script.  If a script has not been saved yet (i.e.
for a new level), the script from the last level is loaded (if there is
one).